+++ src/main/java/com/baichen/rpc/limiter/Limiter.java
@@ -1,0 +1,12 @@
+package com.baichen.rpc.limiter;
+
+public interface Limiter {
+
+    boolean tryAcquire();
+
+    default void release() {
+        release(1);
+    }
+
+    void release(int count);
+}